DAOS-15998 pool: Let all pool_svc rest on ds_pool
Currently, only leader pool_svc objects depend on ds_pool objects. This
unintentionally allows the creation of a PS replica even when the local
ds_pool object is being stopped, leading to the hang in the Jira ticket.
This patch lets every follower pool_svc object depend on the ds_pool
object too, so that if the latter is stopping or doesn't exist, the
pool_svc object won't be created successfully.

Signed-off-by: Li Wei <wei.g.li@intel.com>
Required-githooks: true