DAOS-16927 pool: Fix upgrade DER_NOTSUPPORTEDs
ds_pool_upgrade_if_needed updates the upgrade global version in the RDB TX that
is also used when performing the actual upgrade. If the latter fails, all
updates in the TX are discarded, and the upgrade state becomes:

          upgrade status: FAILED
  upgrade global version: <old version>
          global version: <old version>

If the upgrade is retried, because the upgrade global version is less than the new version, a DER_NOTSUPPORTED will occur.

  - ds_pool_upgrade_if_needed
      - First transit to
          upgrade status: IN_PROGRESS
  upgrade global version: <new version>
          global version: <old version>
      - If fails, transite to
          upgrade status: FAILED
  upgrade global version: <new version>
          global version: <old version>
      - Maintain existing logic
          - Simplify parameters
          - Reorder of out_tx and upgrade function calls
              - Need skip_upgrade for rc == 0 case
      - Use an RDB TX to transit to upgrade state

Signed-off-by: Li Wei <liwei@hpe.com>